Update /userinfo fetch to use instrumented client
We have 2 clients: one with a dynamic client-credentials-grant-issued
bearer token for talking to the Auth0 management API, and another for
"everything else", which in this case, is just this one call.  I set up
the httpClient var to support instrumenting it, but then never updated
the lone request to use it.

Signed-off-by: Eric Searcy <eric@linuxfoundation.org>